// frontend/app/src/components/space/world_graph/types.ts
export type Bounds = {
    left: number;
    right: number;
    top: number;
    bottom: number;
};

export type WorldCanvasGeometry = {
    width: number;
    height: number;
};